Refactored the carouselviewrenderer split all the classes into separate files
