+++ txt/src/carouselView/dev/Shared.Android/Extensions/NumericExtensions.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Xamarin.Forms.Platform.Extensions
+{
+    internal static class NumericExtensions
+    {
+        public static double Clamp(this double self, double min, double max)
+        {
+            return Math.Min(max, Math.Max(self, min));
+        }
+
+        public static int Clamp(this int self, int min, int max)
+        {
+            return Math.Min(max, Math.Max(self, min));
+        }
+    }
+}
